Fix codeclimate and test issues
