+++ app/build.gradle
@@ -7,6 +7,9 @@ jacoco {
 }
 
 android {
+    packagingOptions {
+        exclude 'META-INF/proguard/androidx-annotations.pro'
+    }
     compileSdkVersion 28
     buildToolsVersion "27.0.3"
     defaultConfig {
@@ -76,6 +79,7 @@ dependencies {
         exclude module: 'recyclerview-v7'
     }
     implementation 'org.jetbrains:annotations-java5:15.0'
+    implementation 'androidx.exifinterface:exifinterface:1.0.0'
 }
 
 tasks.withType(Test) {
